split up the config module part 1 - pluginmanager and fs interaction
