+++ testing/test_config.py
@@ -4,7 +4,8 @@ import textwrap
 import pytest
 
 import _pytest._code
-from _pytest.config import getcfg, get_common_ancestor, determine_setup, _iter_rewritable_modules
+from _pytest.config import _iter_rewritable_modules
+from _pytest.config.fs_interaction import getcfg, get_common_ancestor, determine_setup
 from _pytest.main import EXIT_NOTESTSCOLLECTED
 
 
@@ -622,7 +623,7 @@ def test_load_initial_conftest_last_ordering(testdir):
 
 
 def test_get_plugin_specs_as_list():
-    from _pytest.config import _get_plugin_specs_as_list
+    from _pytest.config.pluginmanager import _get_plugin_specs_as_list
     with pytest.raises(pytest.UsageError):
         _get_plugin_specs_as_list(set(['foo']))
     with pytest.raises(pytest.UsageError):
